Extract a CreateNewBoard component and add the buton in the SideBar along with the boards page.
